moved/refactored game types super slightly.
