+++ src/gameData/game1/player/playerShot.ts
@@ -1,4 +1,4 @@
-import type { IEnemyJson } from "../../../App/services/Enemies/enemyConfigs/IEnemyJson";
+import type { IEnemyJson } from "../../../gameTypes/IEnemyJson";
 
 import { spawn } from "../../utils";
