+++ src/gameData/game1/stage2.ts
@@ -1,4 +1,4 @@
-import type { IEnemyJson } from "../../App/services/Enemies/enemyConfigs/IEnemyJson";
+import type { IEnemyJson } from "../../gameTypes/IEnemyJson";
 
 import { attr, forever, moveToAbsolute, parallelAll, setShotSpeed, wait } from "../utils";
 import { col } from "./common";
