+++ src/gameData/game2/enemies/nonShootingAimer.ts
@@ -1,4 +1,4 @@
-import type { IEnemyJson } from "@/App/services/Enemies/enemyConfigs/IEnemyJson";
+import type { IEnemyJson } from "@/gameTypes/IEnemyJson";
 
 import { forever, parallelAll, repeat, setSpeed, spawn, wait } from "@/gameData/utils";
