+++ src/gameData/game2/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-default-export */
-import type { IEnemyJson } from "@/App/services/Enemies/enemyConfigs/IEnemyJson";
+import type { TGame } from "@/gameTypes/TGame";
 
 import { explosion, roundExplosion } from "./effects/explosions";
 import { player } from "./player/player";
@@ -15,7 +15,7 @@ import { spinningDots } from "./enemies/spinningDots";
 import { boss } from "./enemies/boss";
 import { bossCorpse } from "./enemies/bossCorpse";
 
-const game: IEnemyJson[] = [
+const game: TGame = [
    player,
    playerShot,
    parallax,
