+++ src/gameData/game3/pacifistStage.ts
@@ -1,4 +1,4 @@
-import type { IEnemyJson } from "../../App/services/Enemies/enemyConfigs/IEnemyJson";
+import type { IEnemyJson } from "../../gameTypes/IEnemyJson";
 import type { TSpawn } from "../../App/services/Enemies/actions/actionTypes";
 
 import { Do, repeat, wait } from "../utils";
