+++ src/gameTypes/IEnemyJson.ts
@@ -1,5 +1,11 @@
-import type { TAction } from "../actions/actionTypes";
+import type { TAction } from "../App/services/Enemies/actions/actionTypes";
 
+/**
+ * This is the type of a GameObject used when you create a new GameObject.
+ * One can think of this as being a template. When spawning new GameObjects then this "template" is
+ * used to create a new GameObject "instances".
+ * TODO: This type should probably be moved upward.
+ */
 export type IEnemyJson = {
   name: string;
   hp: number; // actually maxHp
