+++ src/gameTypes/README.md
@@ -1,0 +1,3 @@
+These are types that are needed to make a game, i.e. used for auto-completion type-checking
+when making a game/level but some may also be used internally in the app.
+So, maybe the action types should be here too, I dunno.
