+++ src/gameTypes/TGame.ts
@@ -1,0 +1,6 @@
+import type { IEnemyJson } from "./IEnemyJson";
+
+/**
+ * This is the type for a game/level used when you create a level.
+ */
+export type TGame = IEnemyJson[];
